Backend/requirements.txt
# Django e Django REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Banco de dados
psycopg2-binary==2.9.9

# Autenticação e segurança
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Validação e serialização
django-phonenumber-field==7.2.0
phonenumbers==8.13.27

# Configuração de ambiente
python-decouple==3.8
django-environ==0.11.2

# Testes
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
model-bakery==1.17.0

# Desenvolvimento
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Documentação da API
drf-spectacular==0.27.0

# Cache
redis==5.0.1
django-redis==5.4.0

# Celery para tarefas assíncronas
celery==5.3.4

# Formatação e linting
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Utilitários
Pillow==10.2.0
python-dateutil==2.8.2
